com.spotify: return the device ID where playing happens from an action
In fact, it should probably be an optional input parameter rather
than an output parameter, but this works for now.

The goal is two fold. First, it gives us an output parameter,
so the compiler doesn't hit the compatibility paths, and we actually
return something from the action that the agent can talk about.
Second, it is actually nice to say where playback occurs.