Add Sports Skill
In this pull commit, I am adding a new sports skill to the thingpedia common devices in the staging folder. This is the start of my journey creating this skill, and is my very first pull request as well. I have added three files: index.js handles the API and provides data from it, manifest.tt formats how the data will be returned on Almond, and dataset.tt provides examples and possible queries.