Automatically generate unit tests for queries
If a query has no input parameter, and does not already have
a manually-specified unit test, run a basic unit test that checks
that the result type is correct according to the manifest, or
if there is an error, that the error is declared in the manifest.

Also check the type of queries for manually specified tests, so
we don't need to write the redundant checks in the unit tests.