Fix interaction of Home Assistant and cloud sync
Ensure that the Home Assistant gateway device has the right
ownerTier property so it won't be started if it is configured
through cloud sync (which would not be possible because direct
local connection is necessary). Also, move the cloud check to
the start() method instead of the constructor so the device
can be loaded in the cloud (even though it is ineffective)
without being autodeleted.